backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]>=3.2.10,<3.4
httpx==0.28.1
passlib[bcrypt]==1.7.4
pydantic-settings==2.10.1
python-multipart==0.0.20
email-validator==2.2.0
